[Rename] modules/geo
This commit refactors the geo module as part of the Elasticsearch to OpenSearch renaming.
